added a few catch statements to iterators

See merge request Frostdown/PEBCAK!28

+++ core/src/models/Wave.java
@@ -4,11 +4,11 @@ import java.util.ArrayList;
 
 public class Wave {
 	public ArrayList<Enemy> _wave = new ArrayList<Enemy>();
-	public Long waveBegins;
-	public Long waveLength;
+	public Float waveBegins;
+	public Float waveLength;
 	public Long waveSize;
 	
-public Wave(ArrayList<Enemy> w, Long wB, Long wL) {
+public Wave(ArrayList<Enemy> w, Float wB, Float wL) {
 	_wave = w;
 	waveBegins = wB;
 	waveLength = wL;
